Module rust_nickname_generater::templates

All of the templates built into this lib are meant to resemble a litle bit of rust code.

Use get_all_templates() to get them all as a Vec or get_template_by_name() to get a specific one by name.

Current list of example outputs:

username(foo: bar);
username.consume_coffee();
username.count_brain_cells() //0
username.parse::<u64>();
#![allow(username)]
&mut username
&username
#![deny(warnings)] yuna();
*username
Err(username);
username(cookies: Vec<cookie>)
#![github(username)]
Arc<username>
Box<username>
Mutex<username>
!username
Ok(username)
Ok::<(), username>(())
username || something
() -> username
Self::username();
Some("username");
unsafe {username();}
username::MAX
username::MIN
&'a evelyn
assert!("username");
#[cfg(feature = "username"]
dbg!("username")
#![derive(username)]
for x in username
impl Debug for username
impl Display for username
0..=username
username::new();
panic!("username");
poise::username
print!("username");
println!("username");
0..username
serenity::username
std::username
tokio::username
tracing::username
vec![username]

Constants

FUNCTION_TEMPLTES

A list of function looking templates

FUNCTION_VARIABLE_TEMPLTES

A list of templates with the username inside a function

MACRO_TEMPLTES

A list of templates with the username inside a macro

VARIABLE_TEMPLTES

A list of variable looking templates